Upd_060221 * debugmenu project updated - fixed incorrect project settings. * tod_tools project updated - fixed incorrect project settings. * Renamed 'renderer' ->'gfxinternal'. * Class blocks updated - adjusted implementation for 'GetAllocatorType' and 2 unnamed methods. * Clas fragment updated - added implementations for methods 'LoadResourceFile', 'ApplyFragment' and constructor. * Class gfxinternal dx9 vertex updated - added constructor, method 'CreateVertexBuffer', documented used vertex types. * Class gfxinternal dx9 texture updated - added function 'DrawAllTextures', members renamed slightly. * Globals file updated - added utility functions 'CreateUniqueId', 'GetBuildNumberString', 'GetEngineAuthor', 'GetEngineVersionBuild'. * Class node updated - added 'DestroyAddon' & 'GetFragment' methids. * Class restype updated - added 'ApplyFragmentResource' method stub. * Class zipslotinfo updated - renamed members.

+++ TOD_tools/Progress.cpp
@@ -155,7 +155,7 @@ void Progress::UpdateProgressTime(unsigned int unk, UINT64 time)
 	field_24 += unk;
 	if (time >= field_38 || !m_TimeStart)
 	{
-		float v1 = field_24 / field_28;
+		float v1 = (float)(field_24 / field_28);
 		float v2 = field_48 <= NULL ? 0.f : ((time - field_50) / field_48);
 
 		if (v2 > v1)
